get list country done + add unique index name , add id in the CountryView

+++ ApplicationCore/Entities/Address.cs
@@ -10,7 +10,7 @@ namespace ApplicationCore.Entities
         public string Street { get; set; }
         public string BoxNumber { get; set; }
         public string City { get; set; }
-        public short CountryId { get; set; }
+        public long CountryId { get; set; } 
         public bool? Active { get; set; }
 
         public Country Country { get; set; }
